Truncate files if they already exist
As it turns out `Files.write` does not overwrite the file but fills it
with new data up until the end of the binary stream. If the file used to
contain more bytes then the temp file contains the current bytes plus
the end of the previous file. This then causes the Json parser to choke
as there are extra characters at the end of the file.